add sasl plain authentication support
Add a new config option "sasl", which can be set to "none" or "plain".
In none, behavior does not change and will be the same as before, using
the password config option with PASS if present.
When plain is selected, the new "login" option will be coupled with the
existing "password" option to connect to the server with SASL PLAIN
mode.

The implementation currently does blind SASL login. It does not check if
the server supports sasl or sasl plain, and does not verify login errors,
etc. It has been successfully tested with irc.libera.chat.